% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Regression.R
\name{mylnear.Describe}
\alias{mylnear.Describe}
\title{Data Description and Distribution}
\usage{
mylnear.Describe(data)
}
\arguments{
\item{data}{dataframe}
}
\value{
Descriptive tables and Distribution
}
\description{
Data Description and Distribution
}
\examples{
mylnear.Describe(data = mydata)
}
