% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Regression.R
\name{mylnear.Suggest}
\alias{mylnear.Suggest}
\title{Model Suggestion}
\usage{
mylnear.Suggest(model)
}
\arguments{
\item{model}{fitted model}
}
\value{
Suggested Changes
}
\description{
Model Suggestion
}
\examples{
mylnear.Suggest(model = fit)

}
