% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Regression.R
\name{myModel.Diagnose}
\alias{myModel.Diagnose}
\title{Model Diagnosed}
\usage{
myModel.Diagnose(model)
}
\arguments{
\item{model}{fitted model}
}
\value{
plotly plots
}
\description{
Model Diagnosed
}
\examples{
myModel.Diagnose(model = fit)

}
