% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Regression.R
\name{myModel.Predict}
\alias{myModel.Predict}
\title{Predict using a linear regression model}
\usage{
myModel.Predict(model, newdata)
}
\arguments{
\item{model}{Fitted linear model.}

\item{newdata}{data for prediction.}
}
\value{
Predicted values.
}
\description{
Predict using a linear regression model
}
\examples{
predictions <- myModel.Predict(model, newdata = newdata)

}
