% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Regression.R
\name{myModel}
\alias{myModel}
\title{Build Regression Model}
\usage{
myModel(formula, data)
}
\arguments{
\item{formula}{linear regression formula}

\item{data}{Data frame object}
}
\value{
linear model.
}
\description{
Build Regression Model
}
\examples{
fit <- myModel(y ~ x1 + x2, data = mydata)

}
