% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Regression.R
\name{myModel.Report}
\alias{myModel.Report}
\title{Model Report}
\usage{
myModel.Report(model)
}
\arguments{
\item{model}{fitted model}
}
\value{
Reported Values for the model
}
\description{
Model Report
}
