% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Regression.R
\name{myModel.Results}
\alias{myModel.Results}
\title{Tidy Model Results}
\usage{
myModel.Results(model)
}
\arguments{
\item{model}{Linear Model}
}
\value{
as tibble
}
\description{
Tidy Model Results
}
\examples{
myModel.Results(fit)

}
