% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Regression.R
\name{myModel.Significance}
\alias{myModel.Significance}
\title{Test the significance of the linear regression model}
\usage{
myModel.Significance(model)
}
\arguments{
\item{model}{Fitted linear model.}
}
\value{
Model significance.
}
\description{
Test the significance of the linear regression model
}
\examples{
significance <- model_significance(model)

}
