% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Regression.R
\name{myModel.Summary}
\alias{myModel.Summary}
\title{Get a summary of the linear regression model}
\usage{
myModel.Summary(model)
}
\arguments{
\item{model}{Fitted linear model.}
}
\value{
Model summary.
}
\description{
Get a summary of the linear regression model
}
\examples{
summary <- myModel.Summary(model)

}
