% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Regression.R
\name{myRegression}
\alias{myRegression}
\title{This Package Fits a linear model with the response and predictor vars
Dependencies: broom}
\description{
This Package Fits a linear model with the response and predictor vars
Dependencies: broom
}
