% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linear Regression.R
\name{tibble.Predict}
\alias{tibble.Predict}
\title{Tidier Predictions}
\usage{
tibble.Predict(x)
}
\arguments{
\item{x}{list file ran from myModel.Predict to a tidy form}
}
\value{
class tible of the predictions in tidy form
}
\description{
Tidier Predictions
}
\examples{
tidyPred <- tibble.Predict(predictions)
}
